HH-218702 move DataSourcePropertiesStorage to nab-jdbc-common module
