+++ nab-data-source/src/main/java/ru/hh/nab/datasource/DataSourceFactory.java
@@ -31,6 +31,7 @@ import static ru.hh.nab.datasource.DataSourceSettings.USER;
 import ru.hh.nab.datasource.healthcheck.HealthCheckHikariDataSourceFactory;
 import ru.hh.nab.datasource.monitoring.MetricsTrackerFactoryProvider;
 import ru.hh.nab.datasource.monitoring.StatementTimeoutDataSource;
+import ru.hh.nab.jdbc.common.DataSourcePropertiesStorage;
 import ru.hh.nab.jdbc.common.ext.OpenTelemetryJdbcExtension;
 
 public class DataSourceFactory {
