+++ nab-hibernate-persistence-provider/src/main/java/ru/hh/nab/hibernate/datasource/RoutingDataSource.java
@@ -13,12 +13,12 @@ import org.springframework.jdbc.datasource.DelegatingDataSource;
 import org.springframework.jdbc.datasource.lookup.AbstractRoutingDataSource;
 import org.springframework.lang.Nullable;
 import ru.hh.nab.datasource.DataSourceContextUnsafe;
-import ru.hh.nab.datasource.DataSourcePropertiesStorage;
 import static ru.hh.nab.datasource.DataSourceSettings.DATASOURCE_NAME_FORMAT;
 import ru.hh.nab.datasource.DataSourceType;
 import ru.hh.nab.datasource.NamedDataSource;
 import ru.hh.nab.datasource.healthcheck.HealthCheck;
 import ru.hh.nab.datasource.healthcheck.HealthCheckDataSource;
+import ru.hh.nab.jdbc.common.DataSourcePropertiesStorage;
 import ru.hh.nab.jdbc.common.ext.JdbcExtension;
 import ru.hh.nab.metrics.Counters;
 import ru.hh.nab.metrics.StatsDSender;
