+++ nab-hibernate-persistence-provider/src/main/java/ru/hh/nab/hibernate/transaction/DataSourceContextTransactionManager.java
@@ -11,7 +11,7 @@ import org.springframework.transaction.support.DefaultTransactionDefinition;
 import static org.springframework.transaction.support.TransactionSynchronizationManager.isActualTransactionActive;
 import static org.springframework.transaction.support.TransactionSynchronizationManager.isSynchronizationActive;
 import ru.hh.nab.datasource.DataSourceContextUnsafe;
-import ru.hh.nab.datasource.DataSourcePropertiesStorage;
+import ru.hh.nab.jdbc.common.DataSourcePropertiesStorage;
 
 public class DataSourceContextTransactionManager implements PlatformTransactionManager {
