+++ nab-hibernate-persistence-provider/src/main/java/ru/hh/nab/hibernate/transaction/ExecuteOnDataSourceBeanPostProcessor.java
@@ -12,8 +12,8 @@ import org.springframework.beans.BeansException;
 import org.springframework.beans.factory.config.BeanPostProcessor;
 import org.springframework.core.annotation.AnnotationConfigurationException;
 import ru.hh.nab.datasource.DataSourceContextUnsafe;
-import ru.hh.nab.datasource.DataSourcePropertiesStorage;
 import ru.hh.nab.datasource.DataSourceType;
+import ru.hh.nab.jdbc.common.DataSourcePropertiesStorage;
 
 public class ExecuteOnDataSourceBeanPostProcessor implements BeanPostProcessor {
