+++ nab-hibernate-persistence-provider/src/test/java/ru/hh/nab/hibernate/NabSessionFactoryBuilderFactoryTest.java
@@ -33,11 +33,11 @@ import org.springframework.core.io.support.PropertiesLoaderUtils;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import static ru.hh.nab.common.qualifier.NamedQualifier.SERVICE_NAME;
-import ru.hh.nab.datasource.DataSourcePropertiesStorage;
 import ru.hh.nab.datasource.DataSourceType;
 import ru.hh.nab.hibernate.properties.HibernatePropertiesProvider;
 import ru.hh.nab.hibernate.transaction.ExecuteOnDataSource;
 import ru.hh.nab.hibernate.transaction.TransactionalScope;
+import ru.hh.nab.jdbc.common.DataSourcePropertiesStorage;
 import ru.hh.nab.metrics.StatsDSender;
 
 @ExtendWith(SpringExtension.class)
